feature(typescript): converted: appear / stepper
